Implemented templating.  Added html5 boilerplate.
That was fun.  Had to patch Haml!
Not that haml itself was broken, just that the stable version in
Rubygems couldn't recognize sass 3.2.0, which meant that haml dumped
warnings into stderr every time it rendered a scss file.

Beyond that, nothing fancy.  Ruthlessly stole some code from the
reddis-io website to implement a layout/content template scheme.

TODO: There is a POST handler in main.rb that isn't covered by tests.
That's bad.  Its sole purpose is to satisfy my wanting to see POST in
action so at least it's not something that's real functionality.
However it's not useful code, so it shall get cut in the next commit.